Coursework 5, 4B25
 Committer: Junyeob Kim <jk680@cam.ac.uk>

+++ src/boot/ksdk1.1.0/devMMA8451Q.h
@@ -42,7 +42,7 @@
 void		initMMA8451Q(const uint8_t i2cAddress, WarpI2CDeviceState volatile *  deviceStatePointer);
 WarpStatus	readSensorRegisterMMA8451Q(uint8_t deviceRegister, int numberOfBytes);
 WarpStatus	writeSensorRegisterMMA8451Q(uint8_t deviceRegister,
-					uint8_t payloadBtye,
+					uint16_t payloadBtye,
 					uint16_t menuI2cPullupValue);
 WarpStatus	configureSensorMMA8451Q(uint8_t payloadF_SETUP, uint8_t payloadCTRL_REG1, uint16_t menuI2cPullupValue);
 WarpStatus	readSensorSignalMMA8451Q(WarpTypeMask signal,
